Currently under construction
Building functionality for rendering arrays of pixels
